server/requirements.txt
beautifulsoup4==4.4.0
Flask==0.10.1
Flask-Cors==2.1.0
funcsigs==0.4
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
matplotlib==1.4.3
mock==1.3.0
nose==1.3.7
numpy==1.9.2
pbr==1.4.0
Pillow==2.9.0
pymongo==3.0.3
pyparsing==2.0.3
pyproj==1.9.4
pystache==0.5.4
python-dateutil==2.4.2
pytz==2015.4
requests==2.7.0
requests-cache==0.4.10
six==1.9.0
Werkzeug==0.10.4
wheel==0.24.0